requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
faiss-cpu
pypdf
sentence-transformers
httpx
python-dotenv
